refactor: them module
