+++ app.py
@@ -1,8 +1,8 @@
 import tkinter as tk
-from ui.main_window import main_window
+from ui.main_window import MainWindow
 
 
 if __name__ == "__main__":
     root = tk.Tk()
-    app = main_window(root)
+    app = MainWindow(root)
     root.mainloop()
