+++ helper/CustomCombobox.py
@@ -1,4 +1,3 @@
-import tkinter as tk
 from tkinter import ttk
 from tkinter import messagebox
