+++ models/__init__.py
@@ -1,0 +1,2 @@
+from models.attendance.attendance_model import AttendanceModel
+from models.user.user_model import UserModel
