--- models/contract/contract_model.py
@@ -1,11 +1,0 @@
-class contract_model:
-    def __init__(self, contract_id, employee_id, start_date, end_date, create_date, salary, working_hours_per_week, benefits):
-        self.contract_id = contract_id
-        self.employee_id = employee_id
-        self.start_date = start_date
-        self.end_date = end_date
-        self.create_date = create_date
-        self.salary = salary
-        self.benefits = benefits
-
-
