--- models/department/department_model.py
@@ -1,10 +1,0 @@
-class department_model:
-    def __init__(self, department_id, department_name, description, location, manager_id, create_date):
-        self.department_id = department_id
-        self.department_name = department_name
-        self.description = description
-        self.location = location
-        self.manager_id = manager_id
-        self.create_date = create_date
-
-
