--- models/employee/employee_model.py
@@ -1,16 +1,0 @@
-class employee_model:
-    def __init__(self, employee_id, name, date_of_birth, gender, address, phone_number, email, position_id, start_date, id_card_number, password, end_date, username):
-        self.employee_id = employee_id
-        self.name = name
-        self.date_of_birth = date_of_birth
-        self.gender = gender
-        self.address = address
-        self.phone_number = phone_number
-        self.email = email
-        self.position_id = position_id
-        self.start_date = start_date
-        self.id_card_number = id_card_number
-        self.password = password
-        self.end_date = end_date
-        self.username = username
-
