--- models/employee_role/employee_role_model.py
@@ -1,7 +1,0 @@
-class employee_role_model:
-    def __init__(self,employee_role_id, employee_id, role_id):
-        self.employee_role_id = employee_role_id
-        self.employee_id = employee_id
-        self.role_id = role_id
-
-
