--- models/position/position_model.py
@@ -1,9 +1,0 @@
-class position_model:
-    def __init__(self, position_id, position_name, description, department_id, create_date):
-        self.position_id = position_id
-        self.position_name = position_name
-        self.description = description
-        self.department_id = department_id
-        self.create_date = create_date
-
-
