--- models/role/role_model.py
@@ -1,7 +1,0 @@
-class role_model:
-    def __init__(self, role_id, role_name, description):
-        self.role_id = role_id
-        self.role_name = role_name
-        self.description = description
-
-
