+++ service/__init__.py
@@ -1,0 +1,8 @@
+from service.attendance_service import AttendanceService
+from service.contract_service import ContractService
+from service.department_service import DepartMentService
+from service.employee_role_service import EmployeeRoleService
+from service.employee_service import EmployeeService
+from service.license_service import LicenseService
+from service.position_service import PositionService
+from service.role_service import RoleService
