+++ ui/pages/__init__.py
@@ -1,0 +1,10 @@
+from ui.pages.BasePage import BasePage
+from ui.pages.Contract import Contract
+from ui.pages.Department import Department
+from ui.pages.Employee import Employee
+from ui.pages.EmployeeRole import EmployeeRole
+from ui.pages.License import License
+from ui.pages.Overview import Overview
+from ui.pages.Position import Position
+from ui.pages.Role import Role
+from ui.pages.Timesheet import Timesheet
